refactor(ai): split colonial_phase_planner into part-of concern files (#3278)
Splits the 1,904-line monolithic colonial_phase_planner.dart into the
library file (library doc + imports + planColonialPeace) plus three
`part of` concern files, each well below the 1000-physical-line
code-review guideline:

- colonial_phase_planner.dart            (310 lines)  library + peace
- colonial_phase_planner_acquisition.dart (572)       acquisition methods
- colonial_phase_planner_military.dart    (243)       military conquest plan
- colonial_phase_planner_naval.dart       (786)       naval + colonial-lite + civilian

Verbatim move only via `part`/`part of`; all private helpers remain
library-private and accessible across parts, so there is no public-API
or visibility change and behaviour/determinism is preserved exactly.
`part` is the established split pattern in this directory
(expand_phase_planner.dart). Satisfies the #3278 file-split AC for the
colonial planner; expand_phase_planner.dart split deferred to a
follow-up commit on this PR.

Refs #3278
